feat: DishCard

+++ lib/core/models/dimesnion_assignment.dart
@@ -2,14 +2,14 @@ import 'package:equatable/equatable.dart';
 
 class DimensionAssignment extends Equatable {
   final String id;
-  final int price;
+  final double price;
 
   const DimensionAssignment({required this.id, required this.price});
 
   factory DimensionAssignment.fromJson(Map<String, dynamic> json) {
     return DimensionAssignment(
       id: json['id'] as String,
-      price: json['price'] as int,
+      price: json['price'] as double,
     );
   }
